+++ dubbo-remoting/dubbo-remoting-netty4/src/test/java/org/apache/dubbo/remoting/transport/netty4/NettyServerhandler.java
@@ -1,0 +1,5 @@
+package org.apache.dubbo.remoting.transport.netty4;
+
+public class NettyServerhandler {
+
+}
